requirements.txt
# API
fastapi>=0.115
uvicorn[standard]>=0.29
pydantic>=2.0
pydantic-settings>=2.0

# Database
asyncpg>=0.29
sqlalchemy[asyncio]>=2.0
alembic>=1.13

# Cache
redis[hiredis]>=5.0

# Data ingestion
httpx>=0.27
pandas>=2.0
numpy>=1.26
yfinance>=0.2

# Pipeline orchestration
prefect>=3.0

# dbt
dbt-postgres>=1.7

# Utils
python-dotenv>=1.0
tenacity>=8.0        # retry logic for API calls
structlog>=24.0      # structured logging

# Dev
pytest>=8.0
pytest-asyncio>=0.23
httpx                # for API testing
